Added the flex direction #15
I used this style to make the items within the flex container and stacked them vertically by using column value to the flex direction property. It was a little challenge I encounted, by a single click in running the code but resolved it thanks to the freecode camp instructions I've just remembered.