Added the style to the demo first page #7
Sized the container which is like the outer shell of the page. It organizes the style of the elements included such as alignment, padding for spacing around elements, styling of border-raduis for softening the edges and a light green background color.
in the body contain the actual page content and interactive elements .
I used the flex-direction to display the body elements in column and centered the contents to