Added the pre tag style
the pre tag is styled to wrap the text and provide some padding and background color for readability. In fact the overflow is auto on the pre tag ensures that if the Json content is too long, it will still be scrollable. and it will only be scrollable with the pre tag.
 I also changed the font-style of the content to be displayed in the json file in using the pre proporty.