Added the responsive style #16
The media queries are used to enable the design and layout of a website to ajust based on the screen and size, ensuring that content is readable and accessible on a variety of devices, from desketop to smartphones. Otherwise, they help tailor the presentation of content for different types of screens and most importantly they are vital for creating versatile, user-friendly web designs that deliver a consistent and optimized user experience across diverse devices and environments.