Added the nav-links active transition  #16
The change made allows the links to be in the blur and before any hover. It's is active,  in the fact, it can appear very clearly once you're over it, indeed when you leave the sidebar area it goes back to the normal blur style so that you can see something else.